LayeredLayouts

LayeredLayouts.Zarate

LayeredLayouts.add_dummy_nodes!

LayeredLayouts.assign_coordinates

LayeredLayouts.forbid_solution!

LayeredLayouts.force_layers!

LayeredLayouts.layer_by_longest_path_to_source

LayeredLayouts.ordering_problem

LayeredLayouts.solve_positions

LayeredLayouts.try_get

LayeredLayouts.Zarate — Type

Zarate

A Sugiyaqma style layout for DAGs and Sankey diagrams. Based on

Zarate, D. C., Le Bodic, P., Dwyer, T., Gange, G., & Stuckey, P. (2018, April). Optimal sankey diagrams via integer programming. In 2018 IEEE Pacific Visualization Symposium (PacificVis) (pp. 135-139). IEEE.

Fields

time_limit::Dates.Period: how long to spend trying alternative orderings. There are often many possible orderings that have the same amount of crossings. There is a chance that one of these will allow a better arrangement than others. Mostly the first solution tends to be very good. By setting a time_limit > Second(0), multiple will be tried until the time limit is exceeded. (Note: that this is not a maximum time, but rather a limit that once exceeded no more attempts will be made.). If you have a time_limit greater than Second(0) set then the result is no longer deterministic. Note also that this is heavily affected by first call compilation time.

LayeredLayouts.add_dummy_nodes! — Method

add nodes so that no edges span multiple layers, returns mask

LayeredLayouts.assign_coordinates — Method

assign_coordinates(graph, layer2nodes)

Works out the x and y coordinates for each node in the graph. This is via formulating the problem as a QP, and minimizing total distance of links. It maintains the order given in layer2nodess. returns: xs, ys, total_distance

LayeredLayouts.forbid_solution! — Method

forbid_solution!(m, is_before)

Modifies the JuMP model m to forbid the solution present in is_before. is_before must contain solved variables.

LayeredLayouts.force_layers! — Method

Correct the layer of each node, according to the optional parameter by user

LayeredLayouts.layer_by_longest_path_to_source — Method

Calculate the layer of each node

LayeredLayouts.ordering_problem — Method

ordering_problem(::Zarate, graph, layer2nodes; force_order)

Formulates the problem of working out optimal ordering as a MILP.

Returns:

model::Model: the JuMP model that when optized will find the optimal ordering

is_before::AbstractVector{AbstractVector{Variable}}: the variables of the model, which once solved will have value(is_before[n1][n2]) == true if n1 is best arrange before n2.

LayeredLayouts.solve_positions — Method

solve_positions(::Zarate, graph; force_layer, force_order)

Returns:

xs: the xs coordinates of vertices in the layout

ys: the ys coordinates of vertices in the layout

paths: a Dict which, for each edge in graph, contains a Tuple of coordinate vectors (xs, ys).

The layout is calculated on a graph where dummy nodes can be added to the different layers. As a result, plotting edges as straight lines between two nodes can result in more crossings than optimal : edges should instead be routed through these different dummy nodes. paths contains for each edge, a Tuple of vectors, representing that route through the different nodes as x and y coordinates.

Optional arguments:

force_layer: Vector{Pair{Int, Int}} specifies the layer for each node e.g. [3=>1, 5=>5] specifies layer 1 for node 3 and layer 5 to node 5

forceorder: Vector{Pair{Int, Int}} this vector forces the ordering of the nodes in each layer, e.g. `forceorder = [3=>2, 1=>3]` forces node 3 to lay before node 2, and node 1 to lay before node 3

Example:

using Graphs, Plots

g = random_orientation_dag(complete_graph(5))
xs, ys, paths = solve_positions(Zarate(), g)
scatter(xs, ys)
for e in edges(g)
    e_xs, e_ys = paths[e]
    plot!(e_xs, e_ys)
end

LayeredLayouts.try_get — Method

try_get(collection, inds...)

Like getindex, except returns nothing if not found or if the collection itsself is nothing.